+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [5.71.0] - 2024-06-06
+
+* Added `containsRoutes` method to `Nylo` class. Now you can check if a route exists in your app. E.g. `Nylo.containsRoutes(["/home", "/settings"])`
+* Fix `MetroService` duplicating slate file names
+ 
 ## [5.70.0] - 2024-06-05
 
 * Update `NotEmptyRule` Validation rules to include `null`, `Map` and `List` types